refactor TableScanExecutor and IndexScanExecutor (#4089)
