+++ src/coprocessor/dag/executor/limit.rs
@@ -73,17 +73,10 @@ impl<'a> Executor for LimitExecutor<'a> {
 
 #[cfg(test)]
 mod tests {
-    use cop_datatype::FieldTypeTp;
-    use kvproto::kvrpcpb::IsolationLevel;
-    use protobuf::RepeatedField;
-    use tipb::executor::TableScan;
-
     use crate::coprocessor::codec::datum::Datum;
-    use crate::storage::SnapshotStore;
+    use cop_datatype::FieldTypeTp;
 
-    use super::super::scanner::tests::{get_range, new_col_info, TestStore};
-    use super::super::table_scan::TableScanExecutor;
-    use super::super::topn::tests::gen_table_data;
+    use super::super::tests::{gen_table_scan_executor, get_range, new_col_info};
     use super::*;
 
     #[test]
@@ -103,27 +96,18 @@ mod tests {
             vec![Datum::I64(6), Datum::Bytes(b"f".to_vec())],
             vec![Datum::I64(7), Datum::Bytes(b"g".to_vec())],
         ];
-        let table_data = gen_table_data(tid, &cis, &raw_data);
-        let mut test_store = TestStore::new(&table_data);
-        // init table scan meta
-        let mut table_scan = TableScan::new();
-        table_scan.set_table_id(tid);
-        table_scan.set_columns(RepeatedField::from_vec(cis.clone()));
         // prepare range
         let range1 = get_range(tid, 0, 4);
         let range2 = get_range(tid, 5, 10);
         let key_ranges = vec![range1, range2];
-        // init TableScan
-        let (snapshot, start_ts) = test_store.get_snapshot();
-        let store = SnapshotStore::new(snapshot, start_ts, IsolationLevel::SI, true);
-        let ts_ect = TableScanExecutor::new(table_scan, key_ranges, store, false).unwrap();
+        let ts_ect = gen_table_scan_executor(tid, cis, &raw_data, Some(key_ranges));
 
         // init Limit meta
         let mut limit_meta = Limit::default();
         let limit = 5;
         limit_meta.set_limit(limit);
         // init topn executor
-        let mut limit_ect = LimitExecutor::new(limit_meta, Box::new(ts_ect));
+        let mut limit_ect = LimitExecutor::new(limit_meta, ts_ect);
         let mut limit_rows = Vec::with_capacity(limit as usize);
         while let Some(row) = limit_ect.next().unwrap() {
             limit_rows.push(row.take_origin());
